feat: update wishlist product count in header
