+++ src/app/shared/components/ui/card/card-product/card-product.component.ts
@@ -3,7 +3,7 @@ import { Component, InputSignal, inject, input } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { RouteEnum } from '@enum/route.enum';
 import { ProductResponse } from '@models/product/product-response.model';
-import { ProductComponentService } from '@services/components/product-component.service';
+import { WishlistProductComponentService } from '@services/components/wishlist-product-component.service';
 import { ButtonIconSmallComponent } from '@shared/components/ui/buttons/button-icon-small/button-icon-small.component';
 import { PriceDiscountComponent } from '@shared/components/ui/price/price-with-discount/price.component';
 import { RatingComponent } from '@shared/components/ui/rating/rating.component';
@@ -17,7 +17,7 @@ import { ImageModule } from 'primeng/image';
 	styleUrl: './card-product.component.scss',
 })
 export class CardProductComponent {
-	private _productComponentService: ProductComponentService = inject(ProductComponentService);
+	private _wishlistProductComponentService: WishlistProductComponentService = inject(WishlistProductComponentService);
 
 	public product: InputSignal<ProductResponse> = input.required<ProductResponse>();
 	public layout: InputSignal<LayoutTypeInput> = input<LayoutTypeInput>('grid');
@@ -26,6 +26,6 @@ export class CardProductComponent {
 	public routeEnum = RouteEnum;
 
 	public onClick(): void {
-		this._productComponentService.addOrRemoveProductToWishlistAndUpdateSignal(this.product().id);
+		this._wishlistProductComponentService.addOrRemoveProductToWishlistAndUpdateSignal(this.product().id);
 	}
 }
